f/weekly_imports/list_new_emails_for_slack/requirements.txt
rich
slackclient
python-dotenv